+++ next.config.mjs
@@ -4,6 +4,10 @@ const nextConfig = {
   compiler: {
     styledComponents: true,
   },
+  webpack: (config, options) => {
+    config.cache = false;
+    return config;
+  },
 };
 
 export default nextConfig;
