--- public/dummyData.ts
@@ -1,48 +1,0 @@
-import { TestListData } from "./types";
-
-export const DUMMY_DESTINATIONS_DATA_LIST: TestListData = {
-  서울: [
-    {
-      id: 1,
-      destinationName: "경복궁",
-      tags: ["역사", "궁궐", "서울"],
-    },
-    {
-      id: 2,
-      destinationName: "남산",
-      tags: ["산", "전망", "트래킹"],
-    },
-    {
-      id: 3,
-      destinationName: "청계천",
-      tags: ["도심", "산책", "물길"],
-    },
-    {
-      id: 4,
-      destinationName: "북촌 한옥마을",
-      tags: ["역사", "한옥", "서울"],
-    },
-  ],
-  경기: [
-    {
-      id: 5,
-      destinationName: "수원 화성",
-      tags: ["역사", "궁궐", "경기"],
-    },
-    {
-      id: 6,
-      destinationName: "파주 헤이리 예술마을",
-      tags: ["예술", "문화", "경기"],
-    },
-    {
-      id: 7,
-      destinationName: "용인 에버랜드",
-      tags: ["놀이공원", "놀이기구", "경기"],
-    },
-    {
-      id: 8,
-      destinationName: "수원 월미도",
-      tags: ["테마파크", "물놀이", "경기"],
-    },
-  ],
-};
